Move auth endpoints into blueprint.
